Start splitting tmt.cli into submodules
* `tmt.cli` remains the module with utilities and basic code,
* `tmt.cli._root` contains all commands for now, but eventualy it would
  be just the top-level command,
* `tmt status` is moved into `tmt.cli.status` as an example of how it
  would look like.
